backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pillow==10.2.0
opencv-python==4.9.0.80
keras==3.1.1
tensorflow==2.16.1
python-multipart==0.0.9
scikit-learn==1.4.2